Finishing Docker and Prisma setup + creating comment
